Rename style sheet; a few fixes.
